rework post controller
To handle errors correctly the post controller needed some rework. Now
instead of single functions running all around the place, there is one
function collecting the data needed and raising errors when needed.

For that to work, the controller class can now return more classes to
keep the chain going. Unless a class is returned, it will call the
renderer.