add basic middleware handling
This is an initial spike on how the middleware might be included into
the system. It is inspired/copied from GIN, so I still have to work out,
how to make it nicer than it is at the moment.

What is currently missing is Aborting and the addRoute function looks
very ugly.